#183 [refactor-domain] 공연 정보 서비스 -> 공연 서비스 통합, 공연 단건 조회 API, 공연 전체 조회 API ,공연 정보 전체 조회 API 리팩토링
